Add libpcap for receiving and sending ethernet frames.
The goal is to retire KLH10_NET_BPF, _PFLT, _DLPI and _LNX as much as possible.
However where ethernet addresses are manipulated, such operating system dependent methods still need to be used, since libpcap does not handle that.

Code for the other packet filters is still included but most likely does not work.
To use the new code, add -DKLH10_NET_PCAP to the Makefile(s).