Added Linux tap+bridge, plus other cleanups.
However it turns out that Linux bridges are useless, since they freeze
the network when they are actually used (see http://www.microhowto.info/troubleshooting/troubleshooting_ethernet_bridging_on_linux.html#idp86992 )
"The machine appears to freeze
As noted above, adding an interface to a bridge causes it to
stop acting as an Internet Protocol endpoint. This could result
in the machine appearing to freeze.

"The underlying issue is that when an interface is attached to
a bridge then any network addresses need to be bound to the
bridge, not to the interface."

That makes them useless to configure temporarily without disturbing
the rest of the network traffic.